1. ui refactor     a.  revision verifiecation happens in the rool not in each view  item in loop.     b. move fro ui to ui-chakra for custom chakra components     c. move from ui/nanivation into components
