+++ web/src/components/Navbar.tsx
@@ -1,8 +1,8 @@
 import { Box, Group, HStack, Image, LinkBox, Text } from "@chakra-ui/react"
-import Settings from "../settings"
+import Settings from "./chakra-ui/settings"
 import ConnectWallet from "./ConnectWallet"
-import { useColorMode } from "../color-mode"
-import appStore from "../../../store"
+import { useColorMode } from "./chakra-ui/color-mode"
+import appStore from "../store"
 import { useStore } from "zustand"
 import VersionAndDisclaimer from "./VersionAndDisclaimer"
 import { Link } from "react-router-dom"
