+++ web/src/components/chakra-ui/avatar.tsx
